Supprime les fichiers inutiles + Sépare les composants REACT
