--- src/atoms/image/image.css
@@ -1,7 +1,0 @@
-img {
-    width: 200px;
-    height: 268px;
-
-    background-repeat: no-repeat;
-    background-size: cover;
-}
