--- src/atoms/texte/texte.css
@@ -1,32 +1,0 @@
-.type_film {
-    font-style: normal;
-    font-weight: normal;
-    font-size: 16px;
-    line-height: 120%;
-
-    letter-spacing: 0.04em;
-
-    color: #8E8E96;
-}
-
-.titre_film {
-    font-style: normal;
-    font-weight: 600;
-    font-size: 16px;
-    line-height: 130%;
-
-    letter-spacing: 0.01em;
-
-    color: #121212;
-}
-
-.auteur_film {
-    font-style: normal;
-    font-weight: normal;
-    font-size: 16px;
-    line-height: 130%;
-
-    letter-spacing: 0.01em;
-
-    color: #8E8E96;
-}
