--- src/molecules/blocImage/blocImage.css
@@ -1,36 +1,0 @@
-.container_films {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    padding-right: 24px;
-}
-.container_affiche {
-    display: flex;
-    width: 80%;
-}
-
-.images {
-    width: 200px;
-    height: 268px;
-
-    background-repeat: no-repeat;
-    background-size: cover;
-}
-
-.affiche1 {
-    background-image: url(affiche1.png);
-}
-
-.affiche2 {
-    background-image: url(affiche2.png);
-}
-
-.afffiche3 {
-    background-image: url(affiche3.png);
-}
-
-.affiche4 {
-    background-image: url(affiche4.png);
-}
-
