+++ src/molecules/blocImage/blocImage.jsx
@@ -5,12 +5,12 @@ import Image from '../../atoms/image/image'
 
 export default function BlocImage({ src, auteur, type, titre }) {
     return (
-        <div className='container_films'>
-            <div className='affiche1'><Image></Image></div>
+        <div>
+            <div className='container_films'>
+                <Image />
+            </div>
             <div className='textes'>
-                <Texte>
-
-                </Texte>
+                <Texte />
             </div>
         </div>
     )
