+++ src/molecules/contenuSponsorise/contenuSponsorise.jsx
@@ -1,0 +1,9 @@
+function ContenusSponsorises() {
+    return (
+        <div id="contenus-sponsorises">
+            <Article titreArticle="Axa" PhraseDaccroche="Petit rouleur ? (...)" />
+            <Article titreArticle="Oh!MyMag" PhraseDaccroche="Iris Mittenaere officialise sa relation (...)" />
+            <Article titreArticle="Villa Schweppes" PhraseDaccroche="La plus déjantée (...)" />
+        </div>
+    )
+}
