+++ src/molecules/grosTitre/grosTitres.jsx
@@ -1,0 +1,10 @@
+function GrosTitres() {
+    return (
+        <div id="gros-titres">
+            <ArticlePrincipal />
+            <Article src="/image/affiche1.jpg" categorie='Cannes' titre='Parasite' tempsDeLecture={6} />
+            <Article src="/image/affiche2.jpg" categorie='Cannes' titre='Parasite' tempsDeLecture={6} />
+            <Article src="/image/affiche3.jpg" categorie='Cannes' titre='Parasite' tempsDeLecture={6} />
+        </div>
+    )
+}
