+++ src/molecules/navBar/navbar.jsx
@@ -1,0 +1,15 @@
+function NavBar() {
+    return (
+        <div id="navbar">
+            <IconeRecherche />
+            <Section titre='Programme TV' />
+            <Section titre='Cannes 2019' />
+            <Section titre='Cinémas' />
+            <Section titre='Écrans & TV' />
+            <Section titre='Sortir' />
+            <Section titre='Musiques' />
+            <Section titre='Radio & podcasts' />
+            <Section titre='livres' />
+        </div>
+    )
+}
