+++ src/molecules/pub/pub.jsx
@@ -1,0 +1,7 @@
+function Pub() {
+    return (
+        <div id="pub-principale">
+            <Pub />
+        </div>
+    )
+}
